requirements.txt
streamlit
pandas>=2.1.4
numpy
matplotlib.pyplot
seaborn